+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+- Added `stub_workflow_tasks` helper
+
 ## [1.3.0] - 2025-11-09
 
 ### Added